feat: create home and create routes
